requirements.txt
groq>=0.9.0
pydantic>=2.6.0
deepdiff>=7.0.1
python-dotenv>=1.0.0
streamlit>=1.35.0
